# docs/swagger.yaml
definitions:
  model.BlogPost:
    properties:
      body:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/blog-post:
    get:
      consumes:
      - application/json
      description: Retrieve all blog posts in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.BlogPost'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all blog posts
      tags:
      - blog
    post:
      consumes:
      - application/json
      description: Create a new blog post with title and content
      parameters:
      - description: Blog post data
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/model.BlogPost'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.BlogPost'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a blog post
      tags:
      - blog
  /api/v1/blog-post/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a blog post by its unique ID
      parameters:
      - description: Blog post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a blog post
      tags:
      - blog
    get:
      consumes:
      - application/json
      description: Retrieve a single blog post by its unique ID
      parameters:
      - description: Blog post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BlogPost'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a blog post by ID
      tags:
      - blog
    patch:
      consumes:
      - application/json
      description: Update a blog post's title and content by its ID
      parameters:
      - description: Blog post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated blog post data
        in: body
        name: blog
        required: true
        schema:
          $ref: '#/definitions/model.BlogPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BlogPost'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update a blog post
      tags:
      - blog
swagger: "2.0"